Refine TypeScript alias import packaging
