+++ .changeset/typescript-alias-imports.md
@@ -1,0 +1,5 @@
+---
+"@codegraphy-dev/plugin-typescript": minor
+---
+
+Add TypeScript Alias Import relationships from `compilerOptions.paths` so alias-based imports appear as a separate, toggleable edge type in CodeGraphy.
